DOMCore/tests/submissions/Ms2ger/Node-removeChild.html

<!DOCTYPE html>
<title>Node.removeChild</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src=about:blank></iframe>
<script>
test(function() {
  var frameDoc = frames[0].document,
      s = frameDoc.createElement("a")
  assert_equals(s.ownerDocument, frameDoc)
  assert_throws("NOT_FOUND_ERR", function() { document.body.removeChild(s) })
  assert_equals(s.ownerDocument, frameDoc)

  s = frameDoc.createElement("b")
  frameDoc.body.appendChild(s)
  assert_equals(s.ownerDocument, frameDoc)
  assert_throws("NOT_FOUND_ERR", function() { document.body.removeChild(s) })
  assert_equals(s.ownerDocument, frameDoc)

  s = frameDoc.createTextNode("test")
  frameDoc.body.appendChild(s)
  assert_equals(s.ownerDocument, frameDoc)
  assert_throws("NOT_FOUND_ERR", function() { s.removeChild(frameDoc) })
  assert_throws("NOT_FOUND_ERR", function() { s.removeChild(document) })

  assert_throws(new TypeError(), function() { document.body.removeChild(null) })

  assert_throws(new TypeError(), function() { document.body.removeChild({'a':'b'}) })
})
</script>
